Migrate openscap unit tests to pytest
